Refactor authentication and file upload routes into separate modules
